+++ design-system/storybook-addon/src/tokens/components/index.tsx
@@ -1,0 +1,1 @@
+export { default as Tool } from './tool';