+++ help/help/examples/1-Help-Open-Article-Using-Mockup-Api.tsx
@@ -3,7 +3,7 @@ import { AnalyticsListener } from '@atlaskit/analytics-next';
 import Page from '@atlaskit/page';
 
 import Help, { ARTICLE_TYPE } from '../src';
-import type { Article, articleId } from '../src';
+import type { Article, articleId, HistoryItem } from '../src';
 
 import LocaleIntlProvider from '../example-helpers/LocaleIntlProvider';
 import { getArticle } from './utils/mockData';
@@ -21,10 +21,23 @@ const handleEvent = (analyticsEvent: { payload: any; context: any }) => {
 };
 
 const Example: React.FC = () => {
-  const [articleId, setArticleId] = useState<articleId>({
-    id: '1',
-    type: ARTICLE_TYPE.HELP_ARTICLE,
+  const [navigationData, setNavigationData] = useState<{
+    articleId: articleId;
+    history: HistoryItem[];
+  }>({
+    articleId: { id: '1', type: ARTICLE_TYPE.HELP_ARTICLE },
+    history: [],
   });
+
+  const navigationDataSetter = (navigationData: {
+    articleId: articleId;
+    history: HistoryItem[];
+  }): void => {
+    console.log('new navigation data');
+    console.log(navigationData);
+    setNavigationData(navigationData);
+  };
+
   const onGetHelpArticle = (articleId: articleId): Promise<Article> => {
     return new Promise((resolve) => resolve(getArticle(articleId.id)));
   };
@@ -38,8 +51,8 @@ const Example: React.FC = () => {
               <LocaleIntlProvider locale={'en'}>
                 <Help
                   navigation={{
-                    articleId,
-                    articleIdSetter: setArticleId,
+                    navigationData,
+                    setNavigationData: navigationDataSetter,
                   }}
                   helpArticle={{
                     onGetHelpArticle,